[aps,docs] add ln support for lm & att based am decoders
