+++ .gitignore
@@ -1,5 +1,7 @@
 # vscode
 .vscode/
+
+# python
 __pycache__/
 .py[cod]
