+++ CONTRIBUTING.md
@@ -2,10 +2,5 @@
 
 
 1. Following the `setup` step to set up the python environment and git hooks
-
 2. Create a personal fork of the master branch: `git checkout -b <your-name>/<branch-name>`
-
-3. Create the pull request after the your personal branch is done. Remember to
-
-    * use [type hints](https://docs.python.org/3.8/library/typing.html) in the python code
-    * add necessary test cases in [tests](../tests)
+3. Create the pull request after the your personal branch is done. Remember to use [type hints](https://docs.python.org/3.8/library/typing.html) in the python code and add necessary test cases in [tests](../tests)
