+++ docs/issue.md
@@ -1,7 +1,5 @@
-# Known issues
+# Known Issues
 
-1. The recipes under [examples/asr](../examples/asr) are not guaranted to achieve STOA results. Previously the part of the ASR code is wrote for internal dataset.
-2. The code in each module are refactored and optimized several times thus the configuration & recipe examples may be out-of-date.
-3. At the early stage, the APS aims to support jit exporting features which could simplify the cost of model deployment. Currently not all the supported models are tested for this feature.
-4. The PyTorch-based feature extraction (ASR part) are not guaranted to get same (but similar) results as [Kaldi](https://github.com/kaldi-asr/kaldi), but it can be modified easily if you are familiar with the extraction process in Kaldi.
-5. The implementation of the APS's network module (i.e., `aps.asr`, `aps.sse`) is based on author's personal knowledge thus the mismatch between the code and paper may exist.
+1. The configuration & recipe examples may be out-of-date so please follow the commit hash of the git for reproducing.
+2. The PyTorch-based feature extraction (for ASR tasks) are not guaranted to be same (but similar) as [Kaldi](https://github.com/kaldi-asr/kaldi), but it can be modified easily if you are familiar with the extraction process in Kaldi.
+3. The implementation of the APS's network module (i.e., `aps.asr`, `aps.sse`) is based on author's personal knowledge thus the difference between the codes and original papers may exist.
